[#47] 리뷰 반영
