+++ src/main/java/com/flab/makedel/service/PushService.java
@@ -47,18 +47,9 @@ public class PushService {
         }
     }
 
-    public void sendMessageToStandbyRidersInSameArea(String address, PushMessageDTO pushMessage) {
-        Set<String> tokenSet = deliveryDAO.selectStandbyRiderTokenList(address);
-        List<Message> messages = tokenSet.stream().map(token -> Message.builder()
-            .putData("title", pushMessage.getTitle())
-            .putData("content", pushMessage.getContent())
-            .putData("orderReceipt", pushMessage.getOrderReceipt().toString())
-            .putData("createdAt", pushMessage.getCreatedAt())
-            .setToken(token)
-            .build())
-            .collect(Collectors.toList());
-
+    public void sendMessages(List<Message> messages) {
         FirebaseMessaging.getInstance().sendAllAsync(messages);
     }
+
 
 }
